Pulled genesis file parsing into own package, for clarity
