+++ server/commands/start.go
@@ -21,6 +21,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk"
 	"github.com/cosmos/cosmos-sdk/app"
+	"github.com/cosmos/cosmos-sdk/genesis"
 	"github.com/cosmos/cosmos-sdk/version"
 )
 
@@ -119,7 +120,7 @@ func start(rootDir string, basecoinApp *app.BaseApp) error {
 		// If genesis file exists, set key-value options
 		genesisFile := path.Join(rootDir, "genesis.json")
 		if _, err := os.Stat(genesisFile); err == nil {
-			err = app.LoadGenesis(basecoinApp, genesisFile)
+			err = genesis.LoadGenesis(basecoinApp, genesisFile)
 			if err != nil {
 				return errors.Errorf("Error in LoadGenesis: %v\n", err)
 			}
